python/requirements.txt
pynput~=1.7.6
pyserial~=3.5
blessings~=1.7